pyproject.toml
[tool.poetry]
name = "nexus-python"
version = "0.3.0"
description = "A simple interface for interacting with NexusDB."
authors = ["Will Humble <w@astraanalytics.co>"]
license = "MIT"
readme = "README.md"
exclude = ["tests"]

[tool.poetry.dependencies]
python = "^3.11.8"
requests = "^2.25.1"
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
openai = "^1.30.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
